Add tags for users, fixed shared controllers, constraints
