+++ app/controllers/admins_controller.rb
@@ -1,29 +1,17 @@
-class AdminsController < ApplicationController
+class AdminsController < BaseController
+  def add_provider
+    object = Provider.create(:name       => params[:name],
+                             :url        => params[:url],
+                             :token      => params[:token],
+                             :user       => params[:user],
+                             :password   => params[:password],
+                             :verify_ssl => params.fetch(:verify_ssl, true))
+    render json: object.to_hash
+  end
+
   def add_portfolio
     portfolio = Portfolio.create!(portfolio_params)
     render json: portfolio
-  end
-
-  def list_portfolios
-    portfolios = Portfolio.all
-    render json: portfolios
-  end
-
-  def fetch_portfolio_with_id
-    item = Portfolio.find_by(:id => params[:portfolio_id])
-    render json: item
-  end
-
-  def fetch_portfolio_items_with_portfolio
-    portfolio_items = Portfolio.find_by(id: params[:portfolio_id])
-                               .portfolio_items
-    render json: portfolio_items
-  end
-
-  def fetch_portfolio_item_from_portfolio
-    portfolio_item = Portfolio.where(id: params[:portfolio_id], porfolio_item_id: params[:portfolio_item_id])
-                              .includes(:portfolio_items)
-    render json: portfolio_item
   end
 
   def add_portfolio_item_to_portfolio
@@ -36,9 +24,20 @@ class AdminsController < ApplicationController
     render json: portfolio_item
   end
 
-  def list_portfolio_items
-    portfolio_items = PortfolioItem.all
-    render json: portfolio_items
+  def add_to_order
+    order = Order.find(params['order_id'])
+    hash_parameters = []
+    params['parameters'].each do |p|
+      hash_parameters << {:name => p['name'], :value => p['value'], :format => p['format'],
+                          :type => p['type'] }
+    end
+    order_item = OrderItem.create(:order_id    => order.id,
+                                  :catalog_id  => params['catalog_id'],
+                                  :plan_id     => params['plan_id'],
+                                  :provider_id => params['provider_id'],
+                                  :count       => params['count'] || 1,
+                                  :parameters  => hash_parameters)
+    render json: {"item_id" => order_item.id}
   end
 
   private
