+++ config/initializers/load_extensions.rb
@@ -1,2 +1,2 @@
 require 'admins_constraints'
-require 'defaults_constraints'
+require 'users_constraints'
